+++ zipline_app/templatetags/poll_extras.py
@@ -1,5 +1,0 @@
-from django import template
-
-register = template.Library()
-
-register.filter('strftime', lambda value, arg: value.strftime(arg))