+++ zipline_app/tests/zipline_app/test_fill.py
@@ -37,3 +37,10 @@ class FillViewsTests(TestCase):
         response = self.client.get(url)
         self.assertEqual(response.status_code, 200)
 
+    def test_new_fill_zeroqty(self):
+        url = reverse('zipline_app:fills-new')
+        time = '2015-01-01 06:00:00'
+        f1={'pub_date':time,'asset':self.a1a.id,'fill_qty':0,'fill_price':1}
+        response = self.client.post(url,f1)
+        self.assertContains(response,"Quantity 0 is not allowed")
+